[Chore #153114587] Make corrections from pull request review
